+++ src/musikbox/cursespp/MessageQueue.cpp
@@ -82,7 +82,7 @@ void MessageQueue::Remove(IMessageTarget *target, int type) {
 void MessageQueue::Post(IMessagePtr message, int64 delayMs) {
     boost::recursive_mutex::scoped_lock lock(this->queueMutex);
 
-    delayMs = std::max(0LL, delayMs);
+    delayMs = std::max((int64) 0, delayMs);
 
     milliseconds now = duration_cast<milliseconds>(
         system_clock::now().time_since_epoch());
